# statistics: BCE for the histogram